Move slots components outside the Slider
