+++ packages/base/Slider/src/Slider/Slider.tsx
@@ -1,15 +1,12 @@
 import type { ComponentProps } from 'react'
-import React, { forwardRef, useEffect, useRef } from 'react'
+import React, { forwardRef, useRef } from 'react'
 import { Slider as MUIBaseSlider } from '@mui/base/Slider'
-import type {
-  SliderMarkSlotProps,
-  SliderValueLabelSlotProps,
-} from '@mui/base/Slider'
 import { useCombinedRefs, useOnScreen } from '@toptal/picasso-utils'
 import { twJoin } from 'tailwind-merge'
 
-import { getBgColor } from './utils/get-bg-color'
-import { SliderContextProvider, useSliderContext } from './SliderContext'
+import { SliderContextProvider } from './SliderContext'
+import SliderMark from '../SliderMark'
+import SliderValueLabel from '../SliderValueLabel'
 
 type Value = number | number[]
 type ValueLabelDisplay = 'on' | 'auto' | 'off'
@@ -43,85 +40,6 @@ export interface Props extends ComponentProps<typeof MUIBaseSlider> {
   hideThumbOnEmpty?: boolean
   /** Disable track highlight. */
   disableTrackHighlight?: boolean
-}
-
-// We need custom Mark component because we have
-// different bg color based on value of the Slider
-const Mark = ({
-  markActive,
-  ownerState,
-  'data-index': dataIndex,
-  style,
-  hideTrack,
-}: SliderMarkSlotProps & { hideTrack: boolean }) => {
-  return (
-    <span
-      data-index={dataIndex}
-      style={style}
-      className={twJoin(
-        'absolute w-[6px] h-[6px] rounded-[50%] border-[2px] top-[1.5px] border-solid border-white opacity-100 -translate-x-2/4 box-content',
-        getBgColor({ markActive, hideTrack, value: ownerState.value })
-      )}
-    />
-  )
-}
-
-const tooltipStates: Record<ValueLabelDisplay, string> = {
-  off: 'hidden',
-  auto: 'hidden group-hover/thumb:flex justify-center items-center',
-  on: 'flex justify-center items-center',
-}
-
-const ValueLabel = ({
-  children,
-  index = -1,
-  tooltip = 'off',
-  isOnScreen,
-  ownerState,
-}: SliderValueLabelSlotProps & {
-  tooltip: ValueLabelDisplay
-  isOnScreen: boolean
-}) => {
-  const { registerValueLabel, hasTooltipOverflow, checkTooltipsOverlap } =
-    useSliderContext()
-  const ref = useRef<HTMLSpanElement>(null)
-  const sliderValue = ownerState.value
-  const isRangeSlider = Array.isArray(sliderValue)
-  const isRangeSliderCollapsed =
-    isRangeSlider && sliderValue[0] === sliderValue[1]
-
-  const getPosition = () => {
-    if (!hasTooltipOverflow || isRangeSliderCollapsed) {
-      return
-    }
-
-    return index === 0 ? 'right-[calc(100%-13px)]' : 'left-[calc(100%-13px)]'
-  }
-
-  useEffect(() => {
-    if (!isRangeSlider) {
-      return
-    }
-    checkTooltipsOverlap()
-  }, [checkTooltipsOverlap, sliderValue, isRangeSlider])
-
-  useEffect(() => {
-    registerValueLabel(index, ref)
-  }, [index, registerValueLabel])
-
-  return (
-    <span
-      ref={ref}
-      className={twJoin(
-        'absolute will-change-transform shadow-4 transition-transform m-1 text-sm bg-graphite-800 text-white rounded-sm py-[2px] px-2',
-        tooltipStates[tooltip],
-        isOnScreen ? 'bottom-[calc(100%+4px)]' : 'top-[calc(100%+4px)]',
-        getPosition()
-      )}
-    >
-      {children}
-    </span>
-  )
 }
 
 export const Slider = forwardRef<HTMLElement, Props>(function Slider(
@@ -172,8 +90,8 @@ export const Slider = forwardRef<HTMLElement, Props>(function Slider(
           marks={marks}
           disabled={disabled}
           slots={{
-            mark: Mark,
-            valueLabel: ValueLabel,
+            mark: SliderMark,
+            valueLabel: SliderValueLabel,
           }}
           slotProps={{
             mark: {
