+++ packages/base/Slider/src/SliderValueLabel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SliderValueLabel'
